+++ examples/rbac/.wundergraph/wundergraph.server.ts
@@ -1,9 +1,7 @@
 import { configureWunderGraphServer } from '@wundergraph/sdk/server';
-import type { HooksConfig } from './generated/wundergraph.hooks';
-import type { InternalClient } from './generated/wundergraph.internal.client';
 import type { Role } from './generated/wundergraph.server';
 
-export default configureWunderGraphServer<HooksConfig, InternalClient>(() => ({
+export default configureWunderGraphServer(() => ({
 	hooks: {
 		authentication: {
 			/**