+++ packages/react-query/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.31](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.30...@wundergraph/react-query@0.8.31) (2023-05-12)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.30](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.29...@wundergraph/react-query@0.8.30) (2023-05-12)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
 ## [0.8.29](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.28...@wundergraph/react-query@0.8.29) (2023-05-09)
 
 **Note:** Version bump only for package @wundergraph/react-query